Add files via upload
I haven't tested any of this but I have the db connection, some jquery, and the forms. The main thing that needs to be done is using php add the users shows and reviews to their respective rows. You'll see where in my code. I used bootstarp for css but you can add things to the style.css for extra.